% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_null.R
\name{effect.null}
\alias{effect.null}
\title{Specify null effects for multivariate abundance data}
\usage{
effect.null(fit, pred)
}
\arguments{
\item{fit}{objects of class manyglm, typically the result of a call to \link[mvabund]{manyglm}.}

\item{pred}{Name of predictor of interest in quotes.}
}
\description{
`effect.null` returns a coefficient matrix to be parsed to `powersim.manyglm` by default
to specify a null effect.
}
\details{
`effect.null` produces a coefficient matrix with a null effect that is specified by setting the parameter
estimates of a predictor of interest `pred` to 0. This function is used by default in \link[ecopower]{powersim.manyglm}.
Note that intercept values are parameterised as in \link[ecopower]{effect.alt}.
}
\examples{
library(mvabund)
data(spider)
spiddat <- mvabund(spider$abund)
X <- data.frame(spider$x)

#Find null effect size for continuous predictor
glm.spid <- manyglm(spiddat~soil.dry, family="negative.binomial",data=X)
coeff.null <- effect.null(glm.spid,pred="soil.dry")
}
